beetime: add a beeconfig object class
Should be responsible for getting and setting the settings and saving
them throughout sessions, keeping the configuration variables in a sane
state and allowing a more robust upgrade path.

config: add store parameter

Use when setting only one value to avoid having to manually call store()
on the config object.

TODO: Figure out how to assign an instance of the beeconf object to a
top-level object like mw at startup, so we only need one instance? Or
maybe this is not necessary. Figure out if we need a destroy() method,
and what that is supposed to do.