beetime: use conf object & its *set/*get methods
TODO: Maybe it's premature to worry about, but creating 6 instances of
the config object for one sync may be a bit overkill. As noted before,
look into assigning beeconf to a top-level object.

TODO: figure out why `from config import beeconf` fails in lookup.py and
settings.py when located at the top of the file, instead of in the body
of the functions, where I had to put it for now.