beetime: remove premium setting completely
This setting had its origin in me experimenting with PyQt and
had more of a documentation purpose than any real function.
Literally, since it was a redundant way to set the overwrite
switch to True.

TODO: Add a help button to the settings menu to explain
the aggregation settings and maybe at the same time replace
even those settings by a direct "overwrite" toggle, since
even the "aggregation" settings were due to me experimenting
with QComboBoxes.