sync: add seconds to timestamp
This was useful to me for debugging and might help defuse some confusion
with users when using syncing and keeps more data visible as a plus. The
confusion might be due to the total number of datapoints created, which
can get high quickly when using multiple sync settings, e.g. just 3
datapoints at minimum per session when syncing along with AnkiWeb. This
will be an issue at least until I disable uploading of completely
identical datapoints.